Updated binary protocol version to 3.2.71. Added VANILLA_KEYWORD origin. Don't insert selections if limit exceeded. Added lock symobol to completions popup and status bar icon.

+++ src/main/java/com/tabnine/TabNineCompletionContributor.java
@@ -2,6 +2,7 @@ package com.tabnine;
 
 import com.intellij.codeInsight.completion.*;
 import com.intellij.codeInsight.lookup.*;
+import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.util.TextRange;
@@ -9,12 +10,14 @@ import com.tabnine.binary.requests.autocomplete.AutocompleteResponse;
 import com.tabnine.binary.requests.autocomplete.ResultEntry;
 import com.tabnine.general.DependencyContainer;
 import com.tabnine.general.StaticConfig;
+import com.tabnine.lifecycle.BinaryStateService;
 import com.tabnine.prediction.CompletionFacade;
 import com.tabnine.prediction.TabNineCompletion;
 import com.tabnine.prediction.TabNinePrefixMatcher;
 import com.tabnine.prediction.TabNineWeigher;
 import com.tabnine.selections.TabNineLookupListener;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Objects;
@@ -25,6 +28,7 @@ import static com.tabnine.general.Utils.endsWithADot;
 public class TabNineCompletionContributor extends CompletionContributor {
     private final CompletionFacade completionFacade = DependencyContainer.instanceOfCompletionFacade();
     private final TabNineLookupListener tabNineLookupListener = DependencyContainer.instanceOfTabNineLookupListener();
+    private final BinaryStateService binaryStateService = ServiceManager.getService(BinaryStateService.class);
 
     @Override
     public void fillCompletionVariants(@NotNull CompletionParameters parameters, @NotNull CompletionResultSet resultSet) {
@@ -40,6 +44,7 @@ public class TabNineCompletionContributor extends CompletionContributor {
         if (originalMatcher.getPrefix().length() == 0 && completions.results.length == 0) {
             return;
         }
+        this.binaryStateService.limited(completions.is_locked);
 
         resultSet = resultSet.withPrefixMatcher(new TabNinePrefixMatcher(originalMatcher.cloneWithPrefix(completions.old_prefix)))
                 .withRelevanceSorter(CompletionSorter.defaultSorter(parameters, originalMatcher).weigh(new TabNineWeigher()));
@@ -52,19 +57,24 @@ public class TabNineCompletionContributor extends CompletionContributor {
 
     private ArrayList<LookupElement> createCompletions(AutocompleteResponse completions, @NotNull CompletionParameters parameters, @NotNull CompletionResultSet resultSet) {
         ArrayList<LookupElement> elements = new ArrayList<>();
+        final Lookup activeLookup = LookupManager.getActiveLookup(parameters.getEditor());
+        for (int index = 0; index < completions.results.length && index < completionLimit(parameters, resultSet, completions.is_locked); index++) {
+            LookupElement lookupElement = createCompletion(
+                    parameters, resultSet, completions.old_prefix,
+                    completions.results[index], index, completions.is_locked, activeLookup);
 
-        for (int index = 0; index < completions.results.length && index < completionLimit(parameters, resultSet); index++) {
-            LookupElementBuilder lookupElementBuilder = createCompletion(parameters, resultSet, completions.old_prefix, completions.results[index], index);
-
-            if (resultSet.getPrefixMatcher().prefixMatches(lookupElementBuilder)) {
-                elements.add(lookupElementBuilder);
+            if (resultSet.getPrefixMatcher().prefixMatches(lookupElement)) {
+                elements.add(lookupElement);
             }
         }
 
         return elements;
     }
 
-    private int completionLimit(CompletionParameters parameters, CompletionResultSet result) {
+    private int completionLimit(CompletionParameters parameters, CompletionResultSet result, boolean isLocked) {
+        if (isLocked) {
+            return 1;
+        }
         boolean preferTabNine = !endsWithADot(
                 parameters.getEditor().getDocument(),
                 parameters.getOffset() - result.getPrefixMatcher().getPrefix().length()
@@ -74,8 +84,9 @@ public class TabNineCompletionContributor extends CompletionContributor {
     }
 
     @NotNull
-    private LookupElementBuilder createCompletion(CompletionParameters parameters, CompletionResultSet resultSet,
-                                                  String oldPrefix, ResultEntry result, int index) {
+    private LookupElement createCompletion(CompletionParameters parameters, CompletionResultSet resultSet,
+                                                  String oldPrefix, ResultEntry result, int index,
+                                           boolean locked, @Nullable Lookup activeLookup) {
         TabNineCompletion completion = new TabNineCompletion(
                 oldPrefix,
                 result.new_prefix,
@@ -94,30 +105,44 @@ public class TabNineCompletionContributor extends CompletionContributor {
             completion.deprecated = result.deprecated;
         }
 
-        return LookupElementBuilder.create(completion, result.new_prefix)
-                .withInsertHandler((context, item) -> {
-                    int end = context.getTailOffset();
-                    TabNineCompletion lookupElement = (TabNineCompletion) item.getObject();
-                    try {
-                        context.getDocument().insertString(end + lookupElement.oldSuffix.length(), lookupElement.newSuffix);
-                        context.getDocument().deleteString(end, end + lookupElement.oldSuffix.length());
-                    } catch(RuntimeException re) {
-                        Logger.getInstance(getClass()).warn("Error inserting new suffix. End = " + end +
-                                ", old suffix length = " + lookupElement.oldSuffix.length() + ", new suffix length = "
-                                + lookupElement.newSuffix.length(), re);
-                    }
-                }).withRenderer(new LookupElementRenderer<LookupElement>() {
-                    @Override
-                    public void renderElement(LookupElement element, LookupElementPresentation presentation) {
-                        TabNineCompletion lookupElement = (TabNineCompletion) element.getObject();
-
-                        presentation.setTypeText(StaticConfig.BRAND_NAME);
-                        presentation.setItemTextBold(false);
-                        presentation.setStrikeout(lookupElement.deprecated);
-                        presentation.setItemText(lookupElement.newPrefix);
-                        presentation.setIcon(ICON);
-                    }
+        LookupElementBuilder lookupElementBuilder =
+            LookupElementBuilder.create(completion, result.new_prefix)
+            .withRenderer(
+                new LookupElementRenderer<LookupElement>() {
+                  @Override
+                  public void renderElement(
+                      LookupElement element, LookupElementPresentation presentation) {
+                    TabNineCompletion lookupElement = (TabNineCompletion) element.getObject();
+                    final String typeText = (locked ? LIMITATION_SYMBOL : "") + StaticConfig.BRAND_NAME;
+                    presentation.setTypeText(typeText);
+                    presentation.setItemTextBold(false);
+                    presentation.setStrikeout(lookupElement.deprecated);
+                    presentation.setItemText(lookupElement.newPrefix);
+                    presentation.setIcon(ICON);
+                  }
                 });
+        if (locked) {
+            final InsertNothingLookupElement lookupElement = new LimitExceededLookupElement(
+                    lookupElementBuilder, oldPrefix);
+            if (activeLookup != null) {
+                activeLookup.addLookupListener(lookupElement);
+            }
+            return lookupElement;
+        } else {
+            lookupElementBuilder = lookupElementBuilder.withInsertHandler((context, item) -> {
+                int end = context.getTailOffset();
+                TabNineCompletion lookupElement = (TabNineCompletion) item.getObject();
+                try {
+                    context.getDocument().insertString(end + lookupElement.oldSuffix.length(), lookupElement.newSuffix);
+                    context.getDocument().deleteString(end, end + lookupElement.oldSuffix.length());
+                } catch(RuntimeException re) {
+                    Logger.getInstance(getClass()).warn("Error inserting new suffix. End = " + end +
+                            ", old suffix length = " + lookupElement.oldSuffix.length() + ", new suffix length = "
+                            + lookupElement.newSuffix.length(), re);
+                }
+            });
+        }
+        return lookupElementBuilder;
     }
 
     private void addAdvertisement(@NotNull CompletionResultSet result, AutocompleteResponse completions) {
